[ADD DIR] added dir smart_Pointer
